Add Format and Lint to Makefile (#545)
* Add Make commands for format, lint, flake8, spotless, isort, black, and refactor

* Add mypy test

* Add lint tests to CI

* Fix broken Python test

* Fix broken test for Python

* Add black to dependencies

* Remove Python Protos

* Add automatic local linting

* Update precommit names

* Add black exclusions

* Add tensorflow metadata proto generation

* Ignore tf meta directory

* Add build essentials to install make in CI

* Add exports back to __init__.py

* Add __all__ to export

* Add white space to export

* Add source to export

* Fix python export formatting

+++ sdk/python/feast/loaders/file.py
@@ -19,7 +19,7 @@ import tempfile
 import uuid
 from datetime import datetime
 from typing import List, Optional, Tuple, Union
-from urllib.parse import urlparse, ParseResult
+from urllib.parse import ParseResult, urlparse
 
 import pandas as pd
 from google.cloud import storage
@@ -63,13 +63,14 @@ def export_source_to_staging_location(
     # Prepare Avro file to be exported to staging location
     if isinstance(source, pd.DataFrame):
         # DataFrame provided as a source
+        uri_path = None  # type: Optional[str]
         if uri.scheme == "file":
             uri_path = uri.path
-        else:
-            uri_path = None
 
         # Remote gs staging location provided by serving
-        dir_path, file_name, source_path = export_dataframe_to_local(source, uri_path)
+        dir_path, file_name, source_path = export_dataframe_to_local(
+            df=source, dir_path=uri_path
+        )
     elif urlparse(source).scheme in ["", "file"]:
         # Local file provided as a source
         dir_path = None
@@ -108,7 +109,7 @@ def export_source_to_staging_location(
         )
 
     # Clean up, remove local staging file
-    if isinstance(source, pd.DataFrame) and len(str(dir_path)) > 4:
+    if dir_path and isinstance(source, pd.DataFrame) and len(str(dir_path)) > 4:
         shutil.rmtree(dir_path)
 
     return [staging_location_uri.rstrip("/") + "/" + file_name]
@@ -172,7 +173,7 @@ def upload_file_to_gcs(local_path: str, bucket: str, remote_path: str) -> None:
 
         remote_path (str):
             Path within GCS bucket to upload file to, includes file name.
-    
+
     Returns:
         None:
             None
